Add shader registry in CMake
It allows to register predefined shader.
Call register_shader to register the sahder it will add the path to the
sahder in the registry and will create an enum value which can be used
to refer to the shader.
When all shaders are registerd call create_shader_manifest to generate
the header with the paths to predefined shaders and the enum values.